ibam_antiplagiat/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-docx==1.1.2
spacy==3.7.4
fr_core_news_md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
scikit-learn==1.5.0
nltk==3.8.1
numpy==1.26.4
reportlab==4.1.0
cryptography==42.0.7
python-magic==0.4.27
bleach==6.1.0
loguru==0.7.2
pytest==8.2.0
pytest-django==4.8.0
pytest-asyncio==0.23.6
factory-boy==3.3.0
coverage==7.5.3
